Figured out incident PCR numbers
It's important to understand the meaning of patient identification variables in the data, and the relationship between patient identification variables across both datasets. However, after manually checking the datasets we found that the terms "Incident" and "Response" are used inconsistently between files. We fixed the MedStar datasets so that incident PCR  number represents a unique incident/person combination.