Commit before moving person ID
I'm committing the current state of the working version of data_01_clean_medstar. In this version, at line 329 we add a unique person id to the MedStar data that contains DETECT screening responses. At line, 800 we add a unique person id to the MedStar data that contains demographics and health information. Later (1352), we merge the screening data with the demographics and health data, but the person id's don't match. I'm going to create a new version where we create a person id after the merge.