Append MedStar data without DETECT screenings
In this update we separated the MedStar data with DETECT screenings from the MedStar data without DETECT screenings. We then joined the MedStar data with DETECT screenings to matching investigations from APS. Then we append the MedStar data from people that did not recieve a DETECT screening to the merged MedStar/APS data. That last part is the new part.