Added count number of screening items completed
When the DETECT tool was used, it often wasn't the case that all screening items were used. Here, we create a variable that contains the number of screening items completed at each pcr, if any. For this purpose, "Not applicable or No Caregiver" and "Don’t Know" count as completed.