Recode don't know to NA
Previously, in _03_preprocess I modified the DETECT screening variables in place by changing all instances of "Don't Know" or "No Caregiver" to NA. That was problematic when I tried to recreate Table 1. It made it look like each question was skipped more often than it actually was. So, I went back to _03_preprocess, created new variables instead of modifying in place, and pushed the new data through all downstream files.